Removing more redirects
