+++ idaes/models_extra/convergence/heater.py
@@ -15,7 +15,7 @@ __author__ = "John Eslick"
 
 import pyomo.environ as pyo
 from pyomo.network import Arc
-from idaes.power_generation.unit_models.helm import HelmValve
+from idaes.models_extra.power_generation.unit_models.helm import HelmValve
 from idaes.core import FlowsheetBlock, MaterialBalanceType
 from idaes.models.unit_models import Heater
 from idaes.models.properties import iapws95
