+++ idaes/models_extra/convergence/turbine.py
@@ -12,7 +12,7 @@
 #################################################################################
 import pyomo.environ as pyo
 import idaes.core as idaes_core
-from idaes.power_generation.unit_models.helm import HelmIsentropicTurbine
+from idaes.models_extra.power_generation.unit_models.helm import HelmIsentropicTurbine
 import idaes.core.util.convergence.convergence_base as cb
 from idaes.models.properties import iapws95
 from idaes.core.solvers import get_solver
