--- idaes/power_generation/unit_models/__init__.py
@@ -1,52 +1,0 @@
-#################################################################################
-# The Institute for the Design of Advanced Energy Systems Integrated Platform
-# Framework (IDAES IP) was produced under the DOE Institute for the
-# Design of Advanced Energy Systems (IDAES), and is copyright (c) 2018-2021
-# by the software owners: The Regents of the University of California, through
-# Lawrence Berkeley National Laboratory,  National Technology & Engineering
-# Solutions of Sandia, LLC, Carnegie Mellon University, West Virginia University
-# Research Corporation, et al.  All rights reserved.
-#
-# Please see the files COPYRIGHT.md and LICENSE.md for full copyright and
-# license information.
-#################################################################################
-from pyomo.common.deprecation import deprecation_warning
-
-deprecation_warning(
-    "The power_generation.unit_models package has been moved to "
-    "idaes.models_extra.power_generation.unit_models",
-    version="2.0.0.alpha0",
-)
-
-from idaes.models_extra.power_generation.unit_models.feedwater_heater_0D import (
-    FWH0D,
-    FWHCondensing0D,
-)
-from idaes.models_extra.power_generation.unit_models.balance import (
-    BalanceBlockData,
-    BalanceBlock,
-)
-from idaes.models_extra.power_generation.unit_models.boiler_fireside import (
-    BoilerFireside,
-)
-from idaes.models_extra.power_generation.unit_models.boiler_heat_exchanger import (
-    BoilerHeatExchanger,
-)
-from idaes.models_extra.power_generation.unit_models.boiler_heat_exchanger_2D import (
-    HeatExchangerCrossFlow2D_Header,
-)
-from idaes.models_extra.power_generation.unit_models.downcomer import Downcomer
-from idaes.models_extra.power_generation.unit_models.drum import Drum
-from idaes.models_extra.power_generation.unit_models.drum1D import Drum1D
-from idaes.models_extra.power_generation.unit_models.feedwater_heater_0D_dynamic import (
-    FWH0DDynamic,
-)
-from idaes.models_extra.power_generation.unit_models.heat_exchanger_3streams import (
-    HeatExchangerWith3Streams,
-)
-from idaes.models_extra.power_generation.unit_models.steamheater import SteamHeater
-from idaes.models_extra.power_generation.unit_models.waterpipe import WaterPipe
-from idaes.models_extra.power_generation.unit_models.watertank import WaterTank
-from idaes.models_extra.power_generation.unit_models.waterwall_section import (
-    WaterwallSection,
-)
