--- idaes/power_generation/unit_models/helm/__init__.py
@@ -1,50 +1,0 @@
-#################################################################################
-# The Institute for the Design of Advanced Energy Systems Integrated Platform
-# Framework (IDAES IP) was produced under the DOE Institute for the
-# Design of Advanced Energy Systems (IDAES), and is copyright (c) 2018-2021
-# by the software owners: The Regents of the University of California, through
-# Lawrence Berkeley National Laboratory,  National Technology & Engineering
-# Solutions of Sandia, LLC, Carnegie Mellon University, West Virginia University
-# Research Corporation, et al.  All rights reserved.
-#
-# Please see the files COPYRIGHT.md and LICENSE.md for full copyright and
-# license information.
-#################################################################################
-from pyomo.common.deprecation import deprecation_warning
-
-deprecation_warning(
-    "The power_generation.unit_models.helm package has been moved to "
-    "idaes.models_extra.power_generation.unit_models.helm",
-    version="2.0.0.alpha0",
-)
-
-from idaes.models_extra.power_generation.unit_models.helm.splitter import HelmSplitter
-from idaes.models_extra.power_generation.unit_models.helm.valve_steam import (
-    HelmValve,
-    ValveFunctionType,
-)
-from idaes.models_extra.power_generation.unit_models.helm.mixer import HelmMixer
-from idaes.models.unit_models import MomentumMixingType
-from idaes.models_extra.power_generation.unit_models.helm.compressor import (
-    HelmIsentropicCompressor,
-)
-from idaes.models_extra.power_generation.unit_models.helm.turbine import (
-    HelmIsentropicTurbine,
-)
-from idaes.models_extra.power_generation.unit_models.helm.pump import HelmPump
-from idaes.models_extra.power_generation.unit_models.helm.turbine_inlet import (
-    HelmTurbineInletStage,
-)
-from idaes.models_extra.power_generation.unit_models.helm.turbine_stage import (
-    HelmTurbineStage,
-)
-from idaes.models_extra.power_generation.unit_models.helm.turbine_outlet import (
-    HelmTurbineOutletStage,
-)
-from idaes.models_extra.power_generation.unit_models.helm.turbine_multistage import (
-    HelmTurbineMultistage,
-)
-from idaes.models_extra.power_generation.unit_models.helm.condenser_ntu import (
-    HelmNtuCondenser,
-    HelmNtuCondenserData,
-)
